refactor: update import paths for Home and Profile components, remove unused Header.tsx file
