--- src/components/container/layout/Header.tsx
@@ -1,59 +1,0 @@
-import React from 'react';
-import { Link, useNavigate } from 'react-router-dom';
-import { useAuth } from '../../../context/AuthContext';
-
-const Header: React.FC = () => {
-  const { isAuthenticated, user, logout } = useAuth();
-  const navigate = useNavigate();
-
-  const handleLogout = () => {
-    logout();
-    navigate('/login');
-  };
-
-  return (
-    <header className="bg-white shadow-md">
-      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-        <div className="flex justify-between items-center py-6">
-          <div className="flex items-center">
-            <Link to="/" className="text-2xl font-bold text-gray-900">
-              Tea Haven
-            </Link>
-          </div>
-          <nav className="flex items-center space-x-4">
-            <Link to="/" className="text-gray-600 hover:text-gray-900">
-              Home
-            </Link>
-            {!isAuthenticated ? (
-              <>
-                <Link to="/login" className="text-gray-600 hover:text-gray-900">
-                  Login
-                </Link>
-                <Link 
-                  to="/register" 
-                  className="bg-indigo-600 text-white px-4 py-2 rounded-md hover:bg-indigo-700"
-                >
-                  Register
-                </Link>
-              </>
-            ) : (
-              <>
-                <Link to="/profile" className="text-gray-600 hover:text-gray-900">
-                  {user?.username}
-                </Link>
-                <button 
-                  onClick={handleLogout}
-                  className="text-gray-600 hover:text-gray-900"
-                >
-                  Logout
-                </button>
-              </>
-            )}
-          </nav>
-        </div>
-      </div>
-    </header>
-  );
-};
-
-export default Header;
